requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pillow>=10.0
nibabel>=5.1
scikit-image>=0.22
torch>=2.1
torchvision>=0.16
